backend/pyproject.toml
[tool.poetry]
name = "vassal"
version = "0.1.0"
description = "An AI-first productivity tool."
authors = ["J. Anthony Santana <anthony@jasantana.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.1.0"
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
